+++ HISTORY.md
@@ -16,6 +16,9 @@
   `noInlining`, `allowUselessChoice`, and `noAlwaysMatch`.
 * Remove unnecessary failure-reporting checks (T224086)
 * Remove failure checks on rules which always match
+* Implemented Unicode Simple Case Folding algorithm for
+  case-insensitive matches, and the `caselessRestrict`
+  option to modify this for ASCII ranges.
 
 ## 5.0.0 (2025-04-28)
 * Improve function and property documentation